backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
bcrypt
python-dotenv
